CMake: Replace config.cmake.in with file WRITE
The generated meshoptimizerConfig.cmake is almost entirely redundant: it
includes the targets cmake and defines unused macros and functions as
part of the default CMake package config setup.

This is motivated by removing config.cmake.in which is a top-level file
that nobody should ever look at, so this simplifies the file structure a
little bit. See https://blog.vito.nyc/posts/cmake-pkg/ for further
explanation of the redundancy.

It would also be possible to embed the targets directly into the config
but that does not seem necessary and this solution preserves the
existing build setup for distributions.