js: Expose MeshoptSimplifier.generatePositionRemap
meshopt_generatePositionRemap is useful in various simplification
workflows, particularly in clustered simplification to setup
cross-cluster boundary locks, and with permissive simplification to
setup protect bits.

While it's possible to implement this in JS, the efficient
implementations are rather difficult to tune, and the naive
implementation is 40x slower than Wasm. So we expose
generatePositionRemap to fix this.