demo: Switch to MeshoptSimplifier.generatePositionRemap
This makes the entire permissive simplification process ~30% faster on
Firefox, as the inefficient JS remap generation was taken a
disproportionate amount of time.

Also remove ES module shims as this realistically is supported on all
browsers we care about for the purpose of simplification demo, and using
the shims emits redundant JS console messages.