demo: Remove uses of negative cone_weight with buildMeshletsFlex
The original idea has been to change the heuristic to produce higher
quality splits for raytracing focused workflows. Unfortunately, this
doesn't *really* work: the impact on RT efficiency is, on average, net
negative, and in general this doesn't solve the problem very well even
for meshes where it is an improvement. Since a new algorithm,
buildMeshletsSpatial, solves that use case much better, we don't need
this feature anymore.