clusterizer: Remove negative cone_weight from buildMeshletsFlex
The original idea has been to change the heuristic to produce higher
quality splits for raytracing focused workflows. Unfortunately, this
doesn't *really* work: the impact on RT efficiency is, on average, net
negative, and in general this doesn't solve the problem very well even
for meshes where it is an improvement. Since a new algorithm,
buildMeshletsSpatial, solves that use case much better, we don't need
this feature anymore.

Since buildMeshletsFlex is experimental, this can be done without
additional versioning. If we need some sort of adjustment like this in
the future, we can still use negative cone_weight as a flag, it's just
that this specific heuristic does not work very well.