clusterizer: Change KD node representation to store leaf count as is
Instead of storing children count as count-1, we now store it as is in
leaf nodes. In tree nodes, children indicates the size of the left
subtree, which can't be zero by construction. Together, these changes
make it possible to "deactivate" KD nodes by assigning children=0 (with
some traversal tweaks).