clusterizer: Iteratively cut KD branches to speed up traversal
On meshes with many topologically disjoint clusters, kdtreeNearest
spends an increasing amount of time going over already emitted
triangles. To fix this, we can detect leaves that have all triangles
already emitted and flag them as inactive using children=0. This, then,
can be propagated to branch nodes using children=0 as a flag as well.

For branches, we do this *before* visiting either node. This makes
deactivation lazier than it strictly speaking needs to be, but it keeps
the existing recursive traversal structure where the second call is a
tail call, which on some models results in extra speedup.

Overall, while on some meshes this is neutral because the KD tree search
is used rarely, on meshes with more seams this can improve the meshlet
building time anywhere from 15% to 35%.