clusterizer: Simplify meshlet_vertices/triangles memory requirements
With the 4b alignment removed, it is now guaranteed for both
meshlet_vertices and meshlet_triangles that no more than index_count
elements will be filled (for meshlet_triangles it should be precisely
index_count elements...).

This allows us to simplify the documented memory requirements; that
helps when using Flex/Spatial builders with a low min_triangles, where
max_meshlets is quite conservative and usually is a significant
overallocation.