gltfpack: Preserve per-instance colors when using instancing
three.js has supported per-instance colors in their implementations of
EXT_mesh_gpu_instancing via _COLOR_0 attribute; with this change we will
now preserve this attribute if it was present in the input scene.

This attribute is non-standard; however, it can be useful in other
circumstances, for example many CAD scenes use material colors to vary
object colors while using the same, simple, geometry; in the future, it
might be interesting to request that via a separate option to
automatically instance them better.