clusterizer: Improve spatial splits to account for vertex fill
Instead of completely disabling fill heuristic when splitting clusters
that can't fit into max_triangles due to vertex limits, we compute the
number of unique vertices for every split position and use it with the
same formula to bias towards more full splits when fill weight is
higher.

This results in a reasonable SAH-fill tradeoff; it's not as good as our
triangle fill because we don't apply it hierarchically, so it's easy to
run into a situation where a 64-triangle cluster has to be split into
3-4 smaller clusters with suboptimal fill, whereas a higher level split
could produce a 48-triangle cluster which could be split into 2-3.

Still, this has a minimal cost and complexity, and it can noticeably
improve the number of meshlets on some levels without compromising
performance; on Bistro in 64t/64v configuration this reduces cluster
count by 9% at the same RT performance.