clusterizer: Automatically detect available SIMD instruction sets
Similarly to code in vertexfilter/vertexcodec, we now auto-detect and
include SSE2 or NEON headers for future SIMD implementations. This might
be useful in other places in the library in the future so we consolidate
the enablement defines a little further, but the logic is the same as
that in vertexfilter.