clusterizer: Use SSE2 to accelerate boxMerge & boxSurface
Together, boxMerge & boxSurface account for most of the runtime in
bvhSplit and, by extension, buildMeshletsSpatial. This code generally
can't be vectorized automatically in a meaningful way - for example, the
3-float loads can't be extended to 4-float because the compiler doesn't
know if padding is safe to read from. Increasing the size of BVHBox
requires more memory and makes the algorithm less efficient for large
sizes due to bad cache effects.

This change merges boxMerge and boxSurface into a single function to
make it easier to vectorize, and replaces the implementation with SSE2
intrinsics when available. The box array is padded by a single element
to make over-reads safe.

To make sure the writes are fast and safe, BVHBoxT type provides
temporary storage with padded vectors. better memory efficiency This
change makes clusterization 10-25% faster on Zen 4 depending on the
workload size.