clusterizer: Improve buildMeshlets* performance for sparse inputs
When buildMeshlets* functions are used to compute clusterization of a
small subset of a large mesh, memset(used) is the only O(vertex count)
operation and as such can take a disproportionate amount of time.

Similarly to adjacency builds, we automatically detect this case and
only clear the indices that are actually used. On hierarchical
clusterization workloads this makes clusterization ~10% faster in
aggregate.