clusterizer: Simplify and restrict SIMD auto-detection logic
Instead of supporting SIMD code on all 32-bit Intel and ARM targets
under MSVC and all 32-bit ARM targets under gcc/clang, limit this to
cases where SSE2 is guaranteed to work on any compiler, or cases when
AArch64 SIMD is available.

This doesn't affect this code right now (other than disabling it on
32-bit targets), but it means that going forward, we should be free to
use AArch64 specific intrinsics. We don't always have perfect testing
coverage for ARMv7 and everywhere except for decoders, there's no
usecase for SIMD on ARMv7 in the library anyway - so this should make it
easier to copy this SIMD support code to other files in the library in
the future.

MSVC has a gotcha wrt arm_neon.h vs arm_neon64.h, but that was fixed in
VS2019 16.2 so again it's not clear that we should care about this for
the rest of the library.