demo: Specify mesh vertex and attribute data externally
Instead of relying on a hardcoded Vertex structure, we now specify all
mesh inputs externally via a struct. To be able to handle permissive
protect bits, for now we specify a mask that can be used to
automatically generate them from the attribute inputs; in the future we
could additionally support protect bits as an input array.