demo: Improve simplification controls a little further
It's now possible to specify per-level simplification ratio (although
this should rarely if ever be changed), and use permissive mode as a
main mode instead of the fallback - which may improve appearance in
intermediate levels compared to using permissive mode as a fallback.

Also actually return the level 0 clusters via output callback.