demo: Use deindexed mesh subset in simplifySloppy fallback
When using simplifySloppy as a fallback, calling simplifySloppy directly
on the index data similarly to the other calls can be very slow because
it doesn't support sparsity. To work around this we can simply deindex
the subset we are simplifying, and restore the original indices after
simplification - which fixes the performance issues at a very small
cost.

Also, the error returned by the simplifier needs to be manually
converted to absolute units.