simplifier: Improve performance of SimplifySparse on very sparse subsets
In some rare cases of extremely large meshes, e.g. 90M vertices, even
setting a bit per vertex can limit simplification throughput
considerably when simplifying individual clusters in a larger
subdivision.

There is no need to zero out the entire array: we can instead just clear
the bit groups that are referenced by the indices. This should not
change the performance of sparse simplification in general, as the bit
setting loop is more expensive than a sparse clear.

The uncached allocation of a large bit array may still be an issue as
we'd need to pagefault on the indices that *are* referenced - however,
this can be fixed by the caller by providing a per-thread arena that's
sufficiently large to accomodate the bitset.