demo: Support DAG output in clusterlod API
In addition to per-level cluster splits, we need to output cluster and
group membership and bounds to be able to implement a full hierarchical
renderer.

To do this, we output the information via two callbacks; the caller
might need to aggregate this in some cases - for example, each cluster
is part of just one group, but we don't know which group until it's
merged and split later.

Each cluster has the group id ('refined') for the higher quality group
of clusters that created it. Each group has a list of clusters that
belong to it.

If boundsError(groups[cluster.refined]) >= threshold and
boundsError(cluster group) < threshold, the cluster should be rendered.
Leaf groups can be identified because no cluster refers to them via
.refined link.