demo: Remove redundant cluster bounds update
When a group of clusters is merged, simplified and split, the group is
finalized and we will never need to look at these clusters again. Bounds
update was not useful and was potentially confusing, as we'd update
cluster bounds to the bounds of the group they created, as opposed to
the bounds of the group that produced them.

We also no longer need the indices so we can free a little bit of memory
earlier.