demo: Integrate SAH overhead reporting into clustered LOD
Instead of having a separate mode where we only clusterize and compute
SAH overhead, compute SAH overhead for every level of the hierarchy.
This helps understand the impact the hierarchical splits have on RT
performance.

This is fairly expensive and only matters for RT workloads so for now
only compute this when spatial clusterizer is used.