simplifier: Mark meshopt_simplifySloppy as stable
After the addition of vertex locks, no further changes are planned to
the interface (or implementation) of the algorithm; further improvements
are likely to be concentrated in the main simplification algorithm
family instead.

As part of this, add a templated overload for the version with
vertex_lock; when vertex_lock was added, the existing templated overload
was kept to provide source level compatibility, but for consistency we
also need a generic version for the full parameter set.