demo: Add clusterlod.h for clustered LOD implementation
This change reimplements the example code for hierarchical
clusterization in the form of a single header library, that can be used
to generate cluster LOD hierarchies using meshopt_ library functions.

This is intended to be used either directly, or as a starter kit that
you can copy and modify in your engine based on the specific needs.

This can replace the actual implementation in the existing demo in the
future.