partition: Tighten max_partition_size a little further
The 3/8 ratio works reasonably well in practice but it makes it
difficult to hit the maximum power-of-two partition size. For some
applications that use the resulting partitions for workgroup aware
processing, this results in excessive overhead for rare partitions.

Adjust the limit to be target * (1 1/3) instead. This means that with
target=24 the maximum is 32, for example, which makes it easier to hit
the maximum power-of-two. In the future we might expose a flex-style API
for this, but for now this tweak is probably sufficient.