demo: Consolidate callbacks into one combined group+cluster
Instead of calling output_group once followed by N output_cluster calls,
we now call the output callback and provide it with both the group data
and all involved clusters. This makes it easier for the consumer to
aggregate the data in batches.

Instead of numbering the groups internally, the output callback returns
an id that will be returned in future clusters as clodCluster::refined.
That allows the caller to self-manage the DAG relationship as well.