demo: Replace example code in nanite.cpp with clusterlod.h
Instead of a handcrafted implementation that is a little more difficult
to reuse, we now use clusterlod.h implementation as is - with all the
metrics and cut visualization retained.

For cuts, we support error-based cut selection as well as a level-based
one; for metrics, we need to duplicate some code around boundary
computation but otherwise can work more or less off the original data.