demo: Ensure usage of meshopt_ APIs is compatible with 0.xx
To avoid a situation where clusterlod.h is used with meshoptimizer 0.25
or earlier, guard against a couple API details that we're taking
advantage of:

- min_triangles must be a multiple of 4 - meshlet_triangles must have
enough space for padding

This code is still mostly expecting meshoptimizer code that's as late as
this file itself (e.g. to take advantage of sparsity performance
improvements), but now with an earlier version things will just silently
work with slightly worse results.