simplify: Avoid dividing by zero during interpolated search
In some cases, interpolate would be called with arguments that result in
an infinite slope during the computation. This is more or less benign:
interpolate() would return +inf, which would be converted to an integer
(resulting in INT_MAX/INT_MIN depending on hardware), which would then
be clamped to a valid grid size on the next loop iteration.

However, this triggers sanitizer errors and probably not something we
should rely on, so fix this by returning x1 instead (which would
similarly be clamped).