indexgenerator: Mark generatePositionRemap as stable
This algorithm is now used to prepare data for multiple different
algorithms in multiple different contexts (and is used in gltfpack and
clusterlod). It also doesn't really have much room to change as it's
doing a rather specific task and doing it well. So we might as well mark
it as stable early.