partition: Move group bounds into ClusterGroup
Instead of using a separate array for bounds, move this data into
ClusterGroup. This makes it more obvious that the bounds data refers to
updated group bounds as the algorithm is merging groups, and makes it
easier to debug.

These were initially separate to ensure that the cost for bounds is nil
when unused, but since the algorithm is going to use bounds-only merging
as a post-process in the future, the main and preferred mode of
operation will involve included positions and bounds, so it makes less
sense to separate.

The performance effect of this change seems to be below the measurement
noise threshold when bounds are unused regardless.