demo: Add clodConfig::partition_spatial flag
When set (defaults to true), this corresponds to passing the vertex
positions to partitioning algorithm. This matches what we're doing right
now; setting it to false may be useful to understand the impact spatial
grouping has.

Note that right now, meshopt_partitionClusters behavior is similar
between having position data and not; however, that's going to change
soon and the effect will be more meaningful. For best DAG quality it's
recommented to keep this as true, hence it's true in the default
configuration.

Also slightly trim the partition size reservation to reflect the
maximum.