partition: Implement followup spatial partition merging
After we're done with topological merging, we may have groups that are
too small because we only consider topological adjacency when picking
merge candidates.

To avoid downstream issues from the resulting small partitions, we now
automatically merge the resulting groups further - using the same logic
where we don't try to grow groups that are already >= target unless
another group picks them as the target - using spatial partitioning.

For that, we build a tree using spatial splits until the leaves are
reasonably small, and then do pairwise merging within the leaves, using
existing bounds scoring for prioritization.

This significantly improves the partition sizes for meshes with many
disconnected clusters which can naturally result during DAG
construction, and makes it match the target much closer. To disable it,
vertex_positions = NULL can be passed (which will disable spatial
proximity scoring as well but it's much less consequential so pretty
safe to disable).