partition: Reverse the group merging order during spatial merge
For topological merging, because the merged group ends up back in the
heap, the merging process is symmetrical (modulo bounds based scoring
where the order matters a bit). However, since in the spatial merging we
only evaluate each group once as a first group, it's important that
first group gets merged into the second one. That way, we retain the
ability to merge more groups into the same one, which increases our
chances of reaching the target.