js: Remove CJS modules and rename ESM modules to .js
This is an initial commit for the migration towards ESM. meshoptimizer
will stop supporting CJS (except for meshopt_decoder.cjs which will be
added later) to simplify the build and deployment process.

As a first step, remove the CJS files and rename ESM files so that
js/meshopt_clusterizer.js et al is the canonical name for the module.