gltfpack: Introduce support for KHR_meshopt_compression
It's now possible to select the extension used for meshopt compression
via a new -ce argument; -ce ext uses EXT_meshopt_compression (default),
whereas -ce khr switches to KHR_meshopt_compression. This option can be
used to gradually switch the default to KHR in the future.

This option is orthogonal to the compression level, so just using -ce
khr will not compress as much as -cc -ce khr. For the latter there's a
shorthand option -cz.