vertexfilter: Optimize SSE2 quaternion decoding by reducing dependencies
Instead of renormalizing the values (which requires a long latency div)
and then computing w (which requires a long latency sqrt), we can work
in unscaled space, computing w using sqrt based on unscaled values, and
renormalize everything afterwards.

This results in a similar instruction count - we replace 3 muls and one
constant load by one mul and one add - but restructures the dependencies
so that div and sqrt become independent, which results in better port
utilization and improves performance by ~15% on Zen 4/clang.