vertexfilter: Use unscaled quaternion decoding for Wasm SIMD
This does not seem to materially impact performance, at least when using
v8 - however, the performance is inferior to native build before/after
so this is likely running into some other bottlenecks. It is helpful to
keep the SIMD kernels identical so this is still preferable since there's
no performance regression.