vertexfilter: Use unscaled quaternion decoding for scalar version
This mirrors the change in SSE2 decoding flow, where we break up the
dependency chain between division and square root. On scalar version,
this makes the code ~25% faster on Zen 4.

While scalar version should pretty much never be selected in practice so
the performance gains are not meaningful, this makes it easier to debug
and validate the code.