simplifier: Mark meshopt_SimplifyRegularize as stable
This flag works as intended and no significant changes to the
implementation are planned. In the future, at some point we may either
expose granular control of the regularization weight or enable per
vertex regularization flags via locks, but that would need a different
mechanism, so we might as well stabilize this now.