gltfpack: Do not use SharedComponent encoding with -vtf
When texture coordinate bounds are unreasonably large, e.g. if some
triangles have broken UV mapping in a single mesh, this breaks
quantization as the bounds are too large to encode using fixed point.
However, using -vtf in this case doesn't work when -cc is used at the
same time, as it uses the same, large, bounds to establish the
exponents.

The use of SharedComponent predates the introduction of exponent
clamping; with exponent clamping, the extra compression that shared
component provides for texture coordinates is minimal, and as such
should be safe to disable.