gltfpack: Warn if quantization causes texture distortion
For some meshes, quantization based on UV bounds may result in
significant distortion; when fixed point UV coordinates are used, we are
introducing an error that may shift UV charts visibly.

This can be mitigated by using floating-point texture coordinates
(-vtf), but by default some meshes may have obvious texture quality
issues.

Similarly to positions, we now detect that this can happen and emit a
warning, suggesting -vtf or using more bits. Note that unlike positions,
where this computation is relative to mesh extents, here we do this
relative to UV area; if the input uses texture tiling with a high
frequency, quantization is possibly fine because the distortion is not
that significant relative to the area.