simplifier: Reject position updates that increase quadric error
In some cases, the original position is significantly more optimal wrt
positional quadric error compared to the solved position. This can
happen in areas with very complex attribute gradients, where a different
position can in theory lead to better attribute error but in practice
visually isn't an improvement; this can also happen in complex geometric
situations with simpler attributes when volume gradient is oriented in
strange ways which may solve the optimal position to an unpredictable
point.

While we should allow some amount of updates that increase positional
quadric error to allow for better attributes, we now reject updates that
move the position a bit too far. This can result in a substantial
fraction of updates rejected (10%+), but this seems to be beneficial for
the overall appearance.