simplifier: Preserve attribute continuity for Complex vertices
When attributes are solved, each wedge is solved in isolation. This can
create attribute discontinuities for vertices that previously didn't
have them; or, to be more precise, if some attributes were continuous
and some weren't, those that were could become discontinuous after the
update. This is unavoidable and optimal from the quadric point of view:
each wedge has a unique mix of quadrics that targets a different
optimum.

This is mostly okay for seam vertices: because the position doesn't
move, the recomputed texture coordinate is the same, and while the
normals *are* different post-update, there are cases where this
regresses quality and cases where it improves it. From a proper
error-based distance, generally speaking discontinuities are okay.

For complex vertices, however, this can also create UV seams by pulling
the UVs apart for vertices that are on a normal crease, because complex
vertices can have their position updated. This is not always
objectionable from a distance, but is not ideal for closeups.

We now automatically detect this for each attribute and correct the
issue by using the wedge with the maximum weight as the source of
optimal value; this ensures that all wedges solve the same in this case.