simplifier: Split solveQuadrics into separate functions
Now that both position and attribute update is non-trivial, splitting
into two functions makes the code a little easier to digest and easier
to analyze in the profiler.

Also, remove the Kind_Locked check from solveAttributes; de-facto the
behavior is still the same for now, as vertex_update is never set for
locked vertices, but solving attributes on locked vertices is safe
(modulo questions around external locking) and this will be adjusted as
a follow-up.