simplifier: Enable updating attributess for locked vertices
When the vertex is locked through internal topological restrictions,
there isn't a good reason to restrct its attributes from being updated.
Moreover, it's often the case that significant simplification gets all
the way to locked vertices as the terminal state by merging a fair
amount of geometry into them, so the original attributes may get quite
stale and it's valuable to update.

When the vertices are locked externally, the attribute update should be
blocked. This is complicated a little bit by the presence of LockBorder
vs vertex_lock; vertex_lock explicitly locks individual vertices which
is easy to respect, but LockBorder locks vertices on the geometric
border - and vertices on geometric border with complex topology have
Locked classification to begin with, which makes it difficult to
distinguish.

We thus disable the attribute update when vertex_lock is used, but keep
it enabled otherwise; this is likely safe for attributes. For positions,
LockBorder needs to completely disable the update to avoid floating
point drift in finalizeVertices, so that's handled in the function
itself now.

Also, Complex vertex update could sometimes be skipped because both
positions and attributes need vertex_update[remap] to be 1 to update the
vertex; occasionally the root vertex in the complex may not be used in
the final result. We now mark both wedges for the update to prevent
this.