demo: Implement clodLocalIndices helper for clusterlod.h
clusterlod operates on global triangle indices internally; when used for
clustered rendering, typically the data needs to be converted to local
indices (with vertex indirection, or cluster-local vertex data).

Building this data can be done using meshopt_buildMeshletsScan, but it's
not as intuitive *and* requires an internal allocation proportional to
mesh vertex count, which is only fast if thread-local arenas are used;
otherwise the overhead is too large even for good system allocators, and
unacceptable on Windows.

Instead, we can use a simple direct mapped cache based on low index
bits, and scan through the entire meshlet on collisions. For now this is
provided as a clod* function instead of meshopt_ function as the use
case is fairly limited, and an alternative approach could involve saving
the buffers produced by meshopt_buildMeshlets* functions during original
clusterization. As it stands, regenerating local indices costs ~1% of
build time performance.