demo: Implement optional partition sorting in clusterlod.h
When requested, we can now sort partitions using Morton order based on a
representative point inside each partition (which we arbitrarily select
from the last cluster in the partition list). This does not noticeably
affect the resulting DAG, but can be helpful to ensure that the output
groups are more coherent spatially, which can be useful for streaming.