partition: Ensure the recursive merge runs in bounded stack space
While the balance check is generally sufficient to limit the recursion
depth even on odd inputs, it should be possible to construct a
pathological input that computes each subsequent split at ~leaf_size for
one half, which would result in a linear stack depth and a possible
stack overflow.

Instead, once we are over a depth cutoff, switch to halving each
partition; this ensures that the stack depth is logarithmic plus a small
constant.