clusterizer: Ensure the kD tree has a bounded depth
While the balance check is generally sufficient to limit the recursion
depth even on odd inputs, it should be possible to construct a
pathological input that computes each subsequent split at ~leaf_size for
one half, which would result in a linear stack depth and a possible
stack overflow.

Instead we now enforce a strict limit and once at that limit, put the
remaining triangles into a leaf node. This is not necessarily ideal from
traversal performance perspective but is better than running out of
stack, and in practice this seems almost impossible to hit. In the
future we could introduce artificially split nodes for this and the
existing balance check instead, as we only need approximate nearest
queries.