clusterizer: Merge two tail packing conditions together
Neither depth limit nor SAH no-op are likely to be hit on regular
inputs; for better code coverage reporting we can merge them. This
results in more or less similar code flow in practice; we do a little
bit of extra work for the leaf nodes that way, but this condition is
effectively impossible to hit anyway.

Also add a note to spatial order code which is the last recursive
function not covered by comments - the stack use there is bounded as the
splits are ~halving the count each time.