clusterizer: Note that SIMD code over-reads other box
For memory efficiency, we store BVHBox as a flat array of 3-float
vectors; to read the vectors we have to use unaligned SIMD reads. From
the perspective of boxMerge function in isolation, this is unsafe
because the read reads one float past the end of .max vector. However,
because the boxes come from an overallocated array, this is safe in
practice as long as we use the function carefully.