vertexfilter: Switch to vfmaq on ARM64
We currently use vmul/vadd to stay compatible with ARM32, where fma is
not available until vfp4. Since it's unconditionally available on armv8,
and gcc already uses it by default, switching to vfma everywhere
improves consistency and performance.

For simplicity, we stub vfmaq similarly to vdivq/vsqrtq on ARM32, unless
it's supported via VFPV4 (which is off by default, but if it's enabled
we need to avoid redefining a library builtin).