vertexfilter: Use div/sqrt for 16-bit octahedral decode on ARM64
Instead of using rsqrt with Newton-Rhapson refinement, we can use
div/sqrt. Similarly to SSE, this is actually faster than attempting to
correct the refinement result if the instructions are supported, and
makes the filter ~8% faster to decode on Apple M4.

On ARM32, we don't have div/sqrt supported natively, so we keep the
existing sequence as using div/sqrt would require excessive emulation
which would be net slower.