vertexfilter: Use explicit stack store for quaternion decoding on NEON
Using vgetq_lane results in clang lowering the rotates to vector code. This
is less efficient, both because NEON lacks vectorized rotates so it ends
up using excessive shifts and bit ops, and because the loop is heavy on
SIMD instructions and using scalar instructions improves the overlap with
subsequent loop iterations.

This makes quaternion decoding ~25% faster on Apple M4. There is still a
potential for clang to make the wrong choices here during codegen, as the
buffer is not marked as volatile, but for now this is probably good enough.