vertexfilter: Work around Clang vectorizing scalar rotates
The previous workaround with the stack buffer was sufficient for Apple Clang,
but non-Apple Clang versions are different in some way and continue to
generate vectorized rotates. These are too expensive by themselves and also
limit the parallelism that we could get for free by overlapping scalar
instructions with vector ones. For now work around this using volatile.

On Apple Clang, volatile results in some extra stack ops that for some
reason make this function a little faster; but the main goal here is to
avoid rotate vectorization.

On Graviton 3 with Clang 15, this makes quat filter decode ~40% faster.