demo: Add optional error clamping based on edge lengths to clusterlod.h
In some clusters with high attribute divergence, we may simplify the
cluster with an error that leaves triangles small geometrically. Since
the LOD selection is targeting <1px sized triangles, this could result
in the actual triangle density being higher than expected. This should
not be a common problem but might affect some types of content or
setups.

Setting simplify_error_edge_limit to a value around 1 will try to limit
the error so that clusters with overly dense geometry can be switched
more quickly; for example, setting the value to 0.5f will try to adjust
the limit such that triangles are 1-2px on screen even if the attribute
error is higher.

To handle thin triangles we use max(min_edge, max_edge / 2) as a
balance, so that triangles that aren't perfectly equilateral are still
switched earlier but long and thin triangles like wires are preserved.