demo: Update edge visualization to distinguish complex edges on loops
We switch the current (cyan) color to only show complex edges with loop
metadata, similarly to other collapse kinds. For complex-complex edges
that aren't tagged as loops, we instead use a darker magenta.

Both can be collapsed but the non-loop edges usually have a higher error
and will use a slightly different logic in the next commit.