gltfpack: Remove sharp edge protection when using permissive simplification
This should no longer be necessary in most cases to produce reasonable
results; previously, without this collapses along sharp edges would be
penalized in permissive mode but with the parallel collapse
implementation this is no longer an issue. This allows permissive mode
to simplify much further for some types of meshes.