js: Add meshopt_decoder.cjs to package exports
Without this, access to CJS compatibility version of meshopt decoder is
not allowed by node, and potentially other bundlers.

While we could reuse "./decoder" export for this, it seems cleaner to
separate it so that the caller intentionally uses the CJS variant. Node
implicitly supports require on ESM files anyway so it is better that the
exports without a suffix route to the ESM files consistently -
especially because the resulting interface is a little different.