vertexfilter: Adjust encodeFilterColor alpha quantization
Because the reconstruction of alpha involves replicating the low bit
instead of replicating the high bit, for minimum error we should
quantize the value as if it had K bits instead of K-1, and drop the low
bit.

This improves the maximum error by up to 40% depending on the bit count;
the effect on average error is more muted and generally is around 10% or
lower.