gltfpack: Silence misleading warning about EXT/KHR meshopt_compression
When converting between EXT & KHR formats (e.g. using `-ce khr` on a
file with EXT_meshopt_compression or vice versa), the extension present
in the input file is absent in the extensions[] list, which generates a
warning despite the fact that we can correctly decode the data from
either extension.